requirements.txt
fastapi==0.99.1
gradio==3.35.2
langchain==0.0.209
pydantic==1.10.10
python-dotenv==1.0.0
Requests==2.31.0
uvicorn==0.22.0
typing-inspect==0.9.0
typing_extensions==4.7.1
chromadb==0.3.26
openai==0.27.8
dnspython==2.3.0
loguru==0.7.0 
pinecone-client==2.2.2
psycopg2==2.9.6
tqdm==4.65.0
sqlalchemy==2.0.17
